Include shader, reorganized hooks
